Invoke AfterAllCallback only if BeforeAllCallback invoked
Prior to this commit, an AfterAllCallback would always be invoked when
using @TestInstance(Lifecycle.PER_CLASS), even if an exception was
thrown by the test class constructor. This is inappropriate since
implementations of BeforeAllCallback would not be invoked in such
cases, and as a general rule JUnit never invokes AfterX methods or
callbacks if the corresponding BeforeX methods or callbacks were not
invoked.

This commit fixes this issue by ensuring that AfterAllCallbacks are
only invoked if BeforeAllCallbacks were invoked.

Fixes: #1460